use double rather than float for delta

+++ extension/src/beehave_context.cpp
@@ -73,12 +73,12 @@ void BeehaveContext::set_actor(Node* actor)
     this->actor = actor;
 }
 
-float BeehaveContext::get_delta() const
+double BeehaveContext::get_delta() const
 {
     return delta;
 }
 
-void BeehaveContext::set_delta(float delta)
+void BeehaveContext::set_delta(double delta)
 {
     this->delta = delta;
 }
